move EntityType to server actions directory
