--- interfaces/utilities.ts
@@ -1,1 +1,0 @@
-export type EntityType = "books" | "authors"
